# Refs #28459 -- Improved performance of duration expressions on SQLite.